fixing   uninitialised byte

+++ zhuravlev.ivan/I3/finder_same_elements_in_two_strings.cpp
@@ -5,7 +5,7 @@ void finderSameElementsInTwoStrings(const char *first_string, const char *second
 {
   size_t final_size = size_first_string + size_sec_string;
   char tmp_string[final_size];
-  size_t cnt = -1;
+  int cnt = -1;
   for (size_t i = 0; i < size_first_string; ++i)
   {
     for (size_t j = 0; j < size_sec_string; ++j)
@@ -20,12 +20,12 @@ void finderSameElementsInTwoStrings(const char *first_string, const char *second
 
   char result_string[cnt]{};
 
-  for (size_t i = 0; i < cnt; ++i)
+  for (int i = 0; i < cnt; ++i)
   {
     result_string[i] = tmp_string[i];
   }
 
-  for (size_t i = 0; i <= cnt; ++i)
+  for (int i = 0; i <= cnt; ++i)
   {
     std::cout << result_string[i] << " ";
   }
